+++ plugins/diopser/src/editor.rs
@@ -93,11 +93,16 @@ fn top_bar(cx: &mut Context) {
             .top(Pixels(2.0))
             .left(Pixels(8.0))
             .on_mouse_down(|_, _| {
-                // Try to open the Diopser plugin's page when clicking on the title. If this fails
-                // then that's not a problem
-                let result = open::that(Diopser::URL);
-                if cfg!(debug) && result.is_err() {
-                    nih_debug_assert_failure!("Failed to open web browser: {:?}", result);
+                // FIXME: On Windows this blocks, and while this is blocking a timer may proc which
+                //        causes the window state to be mutably borrowed again, resulting in a
+                //        panic. This needs to be fixed in baseview first.
+                if cfg!(not(windows)) {
+                    // Try to open the Diopser plugin's page when clicking on the title. If this
+                    // fails then that's not a problem
+                    let result = open::that(Diopser::URL);
+                    if cfg!(debug) && result.is_err() {
+                        nih_debug_assert_failure!("Failed to open web browser: {:?}", result);
+                    }
                 }
             });
         Label::new(cx, Diopser::VERSION)