+++ plugins/spectral_compressor/src/editor.rs
@@ -115,11 +115,19 @@ fn main_column(cx: &mut Context) {
                     ))])
                     .font_size(30.0)
                     .on_mouse_down(|_, _| {
-                        // Try to open the plugin's page when clicking on the title. If this fails
-                        // then that's not a problem
-                        let result = open::that(SpectralCompressor::URL);
-                        if cfg!(debug) && result.is_err() {
-                            nih_debug_assert_failure!("Failed to open web browser: {:?}", result);
+                        // FIXME: On Windows this blocks, and while this is blocking a timer may
+                        //        proc which causes the window state to be mutably borrowed again,
+                        //        resulting in a panic. This needs to be fixed in baseview first.
+                        if cfg!(not(windows)) {
+                            // Try to open the plugin's page when clicking on the title. If this
+                            // fails then that's not a problem
+                            let result = open::that(SpectralCompressor::URL);
+                            if cfg!(debug) && result.is_err() {
+                                nih_debug_assert_failure!(
+                                    "Failed to open web browser: {:?}",
+                                    result
+                                );
+                            }
                         }
                     });
                 Label::new(cx, SpectralCompressor::VERSION)